Refactored options to CLI helpers #160
