+++ plaso/cli/helpers/__init__.py
@@ -7,16 +7,20 @@ from plaso.cli.helpers import data_location
 from plaso.cli.helpers import dynamic_output
 from plaso.cli.helpers import elastic_output
 from plaso.cli.helpers import event_filters
+from plaso.cli.helpers import extraction
 from plaso.cli.helpers import filter_file
+from plaso.cli.helpers import hashers
 from plaso.cli.helpers import language
 from plaso.cli.helpers import mysql_4n6time_output
 from plaso.cli.helpers import nsrlsvr_analysis
 from plaso.cli.helpers import output_modules
+from plaso.cli.helpers import parsers
 from plaso.cli.helpers import sqlite_4n6time_output
 from plaso.cli.helpers import status_view
 from plaso.cli.helpers import storage_file
 from plaso.cli.helpers import tagging_analysis
 from plaso.cli.helpers import temporary_directory
+from plaso.cli.helpers import text_prepend
 from plaso.cli.helpers import timesketch_output
 from plaso.cli.helpers import viper_analysis
 from plaso.cli.helpers import virustotal_analysis
