+++ plaso/cli/helpers/viper_analysis.py
@@ -86,7 +86,7 @@ class ViperAnalysisArgumentsHelper(interface.ArgumentsHelper):
         options, u'viper_host', default_value=cls._DEFAULT_HOST)
     analysis_plugin.SetHost(host)
 
-    port = cls._ParseIntegerOption(
+    port = cls._ParseNumericOption(
         options, u'viper_port', default_value=cls._DEFAULT_PORT)
     analysis_plugin.SetPort(port)
