+++ tests/cli/extraction_tool.py
@@ -15,90 +15,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
 
   # pylint: disable=protected-access
 
-  _EXPECTED_OUTPUT_EXTRACTION_OPTIONS = u'\n'.join([
-      u'usage: extraction_tool_test.py [--artifact_definitions PATH]',
-      (u'                               [--hashers HASHER_LIST] '
-       u'[--parsers PARSER_LIST]'),
-      u'                               [--preferred_year YEAR] [-p]',
-      u'                               [--process_archives]',
-      (u'                               [--skip_compressed_streams] '
-       u'[--yara_rules PATH]'),
-      u'',
-      u'Test argument parser.',
-      u'',
-      u'optional arguments:',
-      u'  --artifact_definitions PATH, --artifact-definitions PATH',
-      (u'                        Path to a directory containing artifact '
-       u'definitions.'),
-      (u'                        Artifact definitions can be used to '
-       u'describe and'),
-      u'                        quickly collect data data of interest, such as',
-      u'                        specific files or Windows Registry keys.',
-      u'  --hashers HASHER_LIST',
-      (u'                        Define a list of hashers to use by the tool. '
-       u'This is a'),
-      (u'                        comma separated list where each entry is the '
-       u'name of a'),
-      (u'                        hasher, such as "md5,sha256". "all" '
-       u'indicates that all'),
-      (u'                        hashers should be enabled. "none" '
-       u'disables all'),
-      (u'                        hashers. Use "--hashers list" or '
-       u'"--info" to list the'),
-      u'                        available hashers.',
-      u'  --parsers PARSER_LIST',
-      (u'                        Define a list of parsers to use by the tool. '
-       u'This is a'),
-      (u'                        comma separated list where each entry can be '
-       u'either a'),
-      (u'                        name of a parser or a parser list. Each entry '
-       u'can be'),
-      (u'                        prepended with an exclamation mark to negate '
-       u'the'),
-      (u'                        selection (exclude it). The list match is an '
-       u'exact'),
-      (u'                        match while an individual parser matching is '
-       u'a case'),
-      (u'                        insensitive substring match, with support for '
-       u'glob'),
-      (u'                        patterns. Examples would be: "reg" that '
-       u'matches the'),
-      (u'                        substring "reg" in all parser names or the '
-       u'glob'),
-      (u'                        pattern "sky[pd]" that would match all '
-       u'parsers that'),
-      (u'                        have the string "skyp" or "skyd" in its '
-       u'name. All'),
-      (u'                        matching is case insensitive. Use "--parsers '
-       u'list" or'),
-      u'                        "--info" to list the available parsers.',
-      u'  --preferred_year YEAR, --preferred-year YEAR',
-      (u'                        When a format\'s timestamp does not include '
-       u'a year,'),
-      (u'                        e.g. syslog, use this as the initial year '
-       u'instead of'),
-      u'                        attempting auto-detection.',
-      u'  --process_archives, --process-archives',
-      (u'                        Process file entries embedded within archive '
-       u'files,'),
-      (u'                        such as archive.tar and archive.zip. This '
-       u'can make'),
-      u'                        processing significantly slower.',
-      u'  --skip_compressed_streams, --skip-compressed-streams',
-      u'                        Skip processing file content within compressed',
-      u'                        streams, such as syslog.gz and syslog.bz2.',
-      u'  --yara_rules PATH, --yara-rules PATH',
-      (u'                        Path to a file containing Yara rules '
-       u'definitions.'),
-      (u'  -p, --preprocess      Turn on preprocessing. Preprocessing is '
-       u'turned on by'),
-      (u'                        default when parsing image files, however if '
-       u'a mount'),
-      (u'                        point is being parsed then this parameter '
-       u'needs to be'),
-      u'                        set manually.',
-      u''])
-
   _EXPECTED_PERFORMANCE_OPTIONS = u'\n'.join([
       u'usage: extraction_tool_test.py [--buffer_size BUFFER_SIZE]',
       u'                               [--queue_size QUEUE_SIZE]',
@@ -115,23 +31,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
       u'                        (defaults to 125000)',
       u''])
 
-  def testGetParserPresetsInformation(self):
-    """Tests the _GetParserPresetsInformation function."""
-    test_tool = extraction_tool.ExtractionTool()
-
-    parser_presets_information = test_tool._GetParserPresetsInformation()
-    self.assertGreaterEqual(len(parser_presets_information), 1)
-
-    available_parser_names = [name for name, _ in parser_presets_information]
-    self.assertIn(u'linux', available_parser_names)
-
-  def testParseExtractionOptions(self):
-    """Tests the _ParseExtractionOptions function."""
-    test_tool = extraction_tool.ExtractionTool()
-
-    options = test_lib.TestOptions()
-
-    test_tool._ParseExtractionOptions(options)
 
   def testParsePerformanceOptions(self):
     """Tests the _ParsePerformanceOptions function."""
@@ -140,18 +39,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
     options = test_lib.TestOptions()
 
     test_tool._ParsePerformanceOptions(options)
-
-  def testAddExtractionOptions(self):
-    """Tests the AddExtractionOptions function."""
-    argument_parser = argparse.ArgumentParser(
-        prog=u'extraction_tool_test.py', description=u'Test argument parser.',
-        add_help=False, formatter_class=test_lib.SortedArgumentsHelpFormatter)
-
-    test_tool = extraction_tool.ExtractionTool()
-    test_tool.AddExtractionOptions(argument_parser)
-
-    output = self._RunArgparseFormatHelp(argument_parser)
-    self.assertEqual(output, self._EXPECTED_OUTPUT_EXTRACTION_OPTIONS)
 
   def testAddPerformanceOptions(self):
     """Tests the AddPerformanceOptions function."""
@@ -164,83 +51,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
 
     output = self._RunArgparseFormatHelp(argument_parser)
     self.assertEqual(output, self._EXPECTED_PERFORMANCE_OPTIONS)
-
-  def testListHashers(self):
-    """Tests the ListHashers function."""
-    output_writer = test_lib.TestOutputWriter(encoding=u'utf-8')
-    test_tool = extraction_tool.ExtractionTool(output_writer=output_writer)
-
-    test_tool.ListHashers()
-
-    output = output_writer.ReadOutput()
-
-    number_of_tables = 0
-    lines = []
-    for line in output.split(b'\n'):
-      line = line.strip()
-      lines.append(line)
-
-      if line.startswith(b'*****') and line.endswith(b'*****'):
-        number_of_tables += 1
-
-    self.assertIn(u'Hashers', lines[1])
-
-    lines = frozenset(lines)
-
-    self.assertEqual(number_of_tables, 1)
-
-    expected_line = b'md5 : Calculates an MD5 digest hash over input data.'
-    self.assertIn(expected_line, lines)
-
-  def testListParsersAndPlugins(self):
-    """Tests the ListParsersAndPlugins function."""
-    output_writer = test_lib.TestOutputWriter(encoding=u'utf-8')
-    test_tool = extraction_tool.ExtractionTool(output_writer=output_writer)
-
-    test_tool.ListParsersAndPlugins()
-
-    output = output_writer.ReadOutput()
-
-    number_of_tables = 0
-    lines = []
-    for line in output.split(b'\n'):
-      line = line.strip()
-      lines.append(line)
-
-      if line.startswith(b'*****') and line.endswith(b'*****'):
-        number_of_tables += 1
-
-    self.assertIn(u'Parsers', lines[1])
-
-    lines = frozenset(lines)
-
-    self.assertEqual(number_of_tables, 9)
-
-    expected_line = b'filestat : Parser for file system stat information.'
-    self.assertIn(expected_line, lines)
-
-    expected_line = b'bencode_utorrent : Parser for uTorrent bencoded files.'
-    self.assertIn(expected_line, lines)
-
-    expected_line = (
-        b'msie_webcache : Parser for MSIE WebCache ESE database files.')
-    self.assertIn(expected_line, lines)
-
-    expected_line = b'olecf_default : Parser for a generic OLECF item.'
-    self.assertIn(expected_line, lines)
-
-    expected_line = b'plist_default : Parser for plist files.'
-    self.assertIn(expected_line, lines)
-
-    expected_line = (
-        b'chrome_history : Parser for Chrome history SQLite database files.')
-    self.assertIn(expected_line, lines)
-
-    expected_line = b'ssh : Parser for SSH syslog entries.'
-    self.assertIn(expected_line, lines)
-
-    expected_line = b'winreg_default : Parser for Registry data.'
-    self.assertIn(expected_line, lines)
 
 
 if __name__ == '__main__':
